De-dup NHL code.
Closes #896.
