+++ lib/chart_types/waterfall_chart.ts
@@ -77,53 +77,6 @@ export class WaterfallChartInfo extends PlaneChartInfo {
   //   }
   // }
 
-  noticePosted(key: string, value: any) {
-    super.noticePosted(key, value);
-    if (this._store.settings.ui.isNarrativeHighlightEnabled) {
-      if (key === 'utteranceBoundary') {
-        const highlight: Highlight = value;
-        this._prevHighlightNavcode = this._doHighlight(highlight, this._prevHighlightNavcode);
-      } else if (key === 'utteranceEnd') {
-        // So that on the initial transition from auto-narration to manual
-        // span navigation, we don't remove any highlights added in manual mode
-        if (!this._store.paraChart.captionBox.highlightManualOverride) {
-          this._store.clearHighlight();
-          this._store.clearAllSeriesLowlights();
-        }
-        // this._highlightIndex = null;
-        if (this._prevHighlightNavcode) {
-          this.didRemoveHighlight(this._prevHighlightNavcode);
-          this._prevHighlightNavcode = '';
-        }
-      }
-    }
-  }
-
-  protected _doHighlight(highlight: Highlight, prevNavcode: string) {
-    if (highlight.navcode) {
-      if (highlight.navcode.startsWith('series')) {
-        const segments = highlight.navcode.split(/-/);
-        this._store.lowlightOtherSeries(...segments.slice(1));
-      } else {
-        this._store.clearHighlight();
-        this._store.highlight(highlight.navcode);
-        if (prevNavcode) {
-          this.didRemoveHighlight(prevNavcode);
-        }
-        this.didAddHighlight(highlight.navcode);
-      }
-      prevNavcode = highlight.navcode;
-    } else {
-      this._store.clearHighlight();
-      this._store.clearAllSeriesLowlights();
-      if (prevNavcode) {
-        this.didRemoveHighlight(prevNavcode);
-        prevNavcode = '';
-      }
-    }
-    return prevNavcode;
-  }
-
   protected _cumulativeTotalForDatapoint(datapoint: Datapoint): number {
     const series = this._store.model!.atKey(datapoint.seriesKey)!;
     return series.datapoints
